refactor: Push Emojify and avatar loading down to core.ui
This makes them available to (future) code that doesn't inherit from
BaseActivity.
