+++ app/src/main/java/app/pachli/adapter/AccountViewHolder.kt
@@ -18,12 +18,12 @@
 package app.pachli.adapter
 
 import androidx.recyclerview.widget.RecyclerView
-import app.pachli.core.activity.emojify
-import app.pachli.core.activity.loadAvatar
 import app.pachli.core.common.extensions.visible
 import app.pachli.core.designsystem.R as DR
 import app.pachli.core.network.model.TimelineAccount
 import app.pachli.core.ui.LinkListener
+import app.pachli.core.ui.emojify
+import app.pachli.core.ui.loadAvatar
 import app.pachli.databinding.ItemAccountBinding
 import app.pachli.interfaces.AccountActionListener
