+++ app/src/main/java/app/pachli/adapter/PollAdapter.kt
@@ -19,10 +19,10 @@ package app.pachli.adapter
 import android.view.LayoutInflater
 import android.view.ViewGroup
 import androidx.recyclerview.widget.RecyclerView
-import app.pachli.core.activity.emojify
 import app.pachli.core.common.extensions.visible
 import app.pachli.core.network.model.Emoji
 import app.pachli.core.ui.BindingHolder
+import app.pachli.core.ui.emojify
 import app.pachli.core.ui.makeIcon
 import app.pachli.databinding.ItemPollBinding
 import app.pachli.viewdata.PollOptionViewData
