+++ app/src/main/java/app/pachli/components/account/media/AccountMediaGridAdapter.kt
@@ -12,13 +12,13 @@ import androidx.paging.PagingDataAdapter
 import androidx.recyclerview.widget.DiffUtil
 import app.pachli.R
 import app.pachli.adapter.isPlayable
-import app.pachli.core.activity.decodeBlurHash
 import app.pachli.core.common.extensions.show
 import app.pachli.core.common.extensions.visible
 import app.pachli.core.data.model.StatusDisplayOptions
 import app.pachli.core.navigation.AttachmentViewData
 import app.pachli.core.network.model.Attachment
 import app.pachli.core.ui.BindingHolder
+import app.pachli.core.ui.decodeBlurHash
 import app.pachli.databinding.ItemAccountMediaBinding
 import app.pachli.util.getFormattedDescription
 import app.pachli.util.iconResource
