+++ app/src/main/java/app/pachli/components/accountlist/adapter/BlocksAdapter.kt
@@ -18,11 +18,11 @@ package app.pachli.components.accountlist.adapter
 
 import android.view.LayoutInflater
 import android.view.ViewGroup
-import app.pachli.core.activity.emojify
-import app.pachli.core.activity.loadAvatar
 import app.pachli.core.common.extensions.visible
 import app.pachli.core.designsystem.R as DR
 import app.pachli.core.ui.BindingHolder
+import app.pachli.core.ui.emojify
+import app.pachli.core.ui.loadAvatar
 import app.pachli.databinding.ItemBlockedUserBinding
 import app.pachli.interfaces.AccountActionListener
