+++ app/src/main/java/app/pachli/components/compose/ComposeAutoCompleteAdapter.kt
@@ -25,13 +25,13 @@ import android.widget.Filterable
 import androidx.annotation.WorkerThread
 import app.pachli.R
 import app.pachli.core.activity.databinding.ItemAutocompleteAccountBinding
-import app.pachli.core.activity.emojify
-import app.pachli.core.activity.loadAvatar
 import app.pachli.core.common.extensions.visible
 import app.pachli.core.common.util.formatNumber
 import app.pachli.core.designsystem.R as DR
 import app.pachli.core.network.model.Emoji
 import app.pachli.core.network.model.TimelineAccount
+import app.pachli.core.ui.emojify
+import app.pachli.core.ui.loadAvatar
 import app.pachli.databinding.ItemAutocompleteEmojiBinding
 import app.pachli.databinding.ItemAutocompleteHashtagBinding
 import com.bumptech.glide.Glide
