+++ app/src/main/java/app/pachli/components/report/adapter/StatusViewHolder.kt
@@ -21,7 +21,6 @@ import android.view.View
 import androidx.recyclerview.widget.RecyclerView
 import app.pachli.R
 import app.pachli.components.report.model.StatusViewState
-import app.pachli.core.activity.emojify
 import app.pachli.core.common.extensions.hide
 import app.pachli.core.common.extensions.show
 import app.pachli.core.common.util.AbsoluteTimeFormatter
@@ -35,6 +34,7 @@ import app.pachli.core.network.model.Status
 import app.pachli.core.network.parseAsMastodonHtml
 import app.pachli.core.ui.LinkListener
 import app.pachli.core.ui.SetStatusContent
+import app.pachli.core.ui.emojify
 import app.pachli.core.ui.setClickableMentions
 import app.pachli.databinding.ItemReportStatusBinding
 import app.pachli.util.StatusViewHelper
