+++ core/ui/build.gradle.kts
@@ -31,9 +31,8 @@ android {
 }
 
 dependencies {
-    // Calls "openLink" from projects.core.activity
-    implementation(projects.core.activity)
     implementation(projects.core.common)
+    implementation(projects.core.data)
     implementation(projects.core.designsystem)
     implementation(projects.core.preferences)
         ?.because("PreferenceEnum types in EnumListPreference")
@@ -42,6 +41,9 @@ dependencies {
         ?.because("Uses HttpException")
     implementation(projects.core.network)
         ?.because("ThrowableExtensions uses getServerErrorMessage")
+
+    implementation(libs.bundles.glide)
+        ?.because("Loads account avatars and emojis")
 
     // Uses JsonDataException from Moshi
     implementation(libs.moshi)
